Rename cdb to crdb.
